rename ops to cli
