+++ pkg/cli/dispute.go
@@ -1,7 +1,7 @@
 // Copyright (c) The Tellor Authors.
 // Licensed under the MIT License.
 
-package ops
+package cli
 
 import (
 	"context"
